% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{nlcor}
\alias{nlcor}
\title{Compute Nonlinear Correlation}
\usage{
nlcor(x, y, refine = 0.5, plt = F)
}
\arguments{
\item{x}{A numeric vector. NAs are not allowed. The length of the vector should be more than 10.}

\item{y}{A numeric vector. NAs are not allowed. Length should be same as x.}

\item{refine}{Optional. Default value 0.5. Increase the value to increase
the granularity of local correlation computation.}

\item{plt}{Optional. Default value FALSE. Set TRUE to return ggplot2 object
for the data correlation visualization.}
}
\value{
The output is a list containing the nonlinear correlation \code{cor.estimate}, \code{adjusted.p.value}, and \code{cor.plot}.
}
\description{
Compute Nonlinear Correlation
}
\details{
Compute nonlinear correlation using spatial sampling. Local linear
correlations are computed at the samples and combined.
}
\section{Functions}{
\itemize{
\item \code{nlcor}: The estimate is between 0 and 1. An adjusted \code{p.value} is also returned.
}}

\examples{
library(nlcor)
plot(x1, y1)
c <- nlcor(x1, y1)
c
c <- nlcor(x2, y2, plt = T)
print(c$cor.plot)
c <- nlcor(x3, y3, refine = 0.9, plt = T)
c$cor.estimate
c$adjusted.p.value
print(c$cor.plot)

}
\seealso{
\code{\link{cor}}
}
\keyword{correlation}
\keyword{nonlinear}
