% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{FindSegmentSize}
\alias{FindSegmentSize}
\title{Find the segmentation size.}
\usage{
FindSegmentSize(l, refine)
}
\arguments{
\item{l}{An integer.}
}
\value{
A floating number \code{s} between 0 and 1.
}
\description{
Identify a segment size that yields at least 5 record.
}
\examples{
FindSegmentSize(l = 100, refine = 0.9)

}
