% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{NetCor}
\alias{NetCor}
\title{Find net correlation from the "linear" segments.}
\usage{
NetCor(cors, pvalues, p.threshold = 0.05)
}
\arguments{
\item{cors}{A vector correlations.}

\item{pvalues}{A vector of pvalues corresponding to the correlations in the \code{cors} vector.}

\item{p.threshold}{The overall threshold of p value, also known as the significance level.}
}
\value{
The net correlation estimate, \code{cor.estimate}, and a list containing
the adjusted correlations and pvalues for each "linear" segment.
}
\description{
Find net correlation from the "linear" segments.
}
\examples{
cors <- c(-0.70, 0.93, -0.79, 0.91)
pvalues <- c(0.004, 0.0006, 0.0007, 0.009)
NetCor(cors, pvalues)

}
