% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlCor.R
\name{nlCor}
\alias{nlCor}
\title{Non Linear Correaltion function - Correlation coeffecient by considering non linear realtionships among two variables}
\usage{
nlCor(X, Y, inverse = F, n = 2, method = "log", na.rm = F)
}
\arguments{
\item{X}{: first dataset (can be a matrix or dataframe or array)}

\item{Y}{: second dataset (can be a matrix or dataframe or array)}

\item{inverse}{: is to take the inverse relation eg. checking 1/X^2,1/X^3 insted of X^2, X^3}

\item{n}{: Maximum Order number you would want to check}

\item{method}{: log(logarithmic), poly(Polynomial), expo(Exponential), Linear. Default is log}
}
\value{
nlCor(X,Y)
}
\description{
Non Linear Correaltion function - Correlation coeffecient by considering non linear realtionships among two variables
}
\examples{
X <- mtcars[,1:2]
Y <- mtcars[,1:2]
nlCor(X,Y,method="poly",n=3)
}
