% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{PlotData}
\alias{PlotData}
\title{Plotting the the scatter plot of x,y}
\usage{
PlotData(x, y)
}
\arguments{
\item{x}{A numeric vector. NAs are not allowed.}

\item{y}{A numeric vector. NAs are not allowed. Length should be same as x.}
}
\value{
ggplot plot object
}
\description{
Scatter plot of x and y
}
\examples{
library(ggplot2)
PlotData(x, y)

}
\keyword{plot}
