% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{SampleCor}
\alias{SampleCor}
\title{Correlation from spatial sampling.}
\usage{
SampleCor(x, y, s)
}
\arguments{
\item{x}{A numeric vector. NAs are not allowed.}

\item{y}{A numeric vector. NAs are not allowed. Length should be same as x.}

\item{s}{The sample size as percent of the vector length. A float number between 0 and 1.}
}
\value{
\code{list(cor, pvalue)} containing the correlations and its pvalue for each segment.
}
\description{
Compute nonlinear correlation from local linear correlations
at some spatial sampling
}
\examples{
SampleCor(x, y, s = 0.2)

}
\keyword{correlation,}
\keyword{nonlinear,}
\keyword{sample}
