% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{Segment}
\alias{Segment}
\title{Segmentation.}
\usage{
Segment(l, s)
}
\arguments{
\item{l}{An integer.}

\item{s}{A float number between 0 and 1.}
}
\value{
The list of segmented sequence \code{1:l}.
}
\description{
Segment \code{1:l} in to almost equal set of indices.
}
\examples{
Segment(l = 100, s = 0.2)

}
