\name{x}
\alias{x}
\docType{data}
\title{
x
}
\description{
a simple vector of 48 observation. This data set should be used with y to see the correlation
}
\usage{data("x")}
\format{
  The format is:
 num [1:48] 30.729 11.286 33.706 34.347 -0.698 ...
}
\source{
It is a simulated data. It is not from any specific source
}
\keyword{datasets}
