\name{x1}
\alias{x1}
\docType{data}
\title{
x1
}
\description{
a simple vector of 1095 observation. This data set should be used with y1 to see the correlation
}
\usage{data("x1")}
\format{
  The format is:
 int [1:1095] 1 2 3 4 5 6 7 8 9 10 ...
}
\source{
It is a simulated data. It is not from any specific source
}
\keyword{datasets}
