\name{x2}
\alias{x2}
\docType{data}
\title{
x2
}
\description{
a simple vector of 701 observation. This data set should be used with y2 to see the correlation
}
\usage{data("x2")}
\format{
  The format is:
 num [1:701] -3 -2.99 -2.98 -2.97 -2.96 -2.95 -2.94 -2.93 -2.92 -2.91 ...
}
\source{
It is a simulated data. It is not from any specific source
}
\keyword{datasets}
