\name{x3}
\alias{x3}
\docType{data}
\title{
x3
}
\description{
a simple vector of 4001 observation. This data set should be used with y3 to see the correlation
}
\usage{data("x3")}
\format{
  The format is:
 num [1:4001] -6.28 -6.28 -6.28 -6.27 -6.27 ...
}
\source{
It is a simulated data. It is not from any specific source
}
\keyword{datasets}
