\name{y}
\alias{y}
\docType{data}
\title{
y
}
\description{
a simple vector of 48 observation. This data set should be used with x to see the correlation
}
\usage{data("y")}
\format{
  The format is:
 num [1:48] 5.648 0.515 5.636 2.291 5.261 ...
}
\source{
It is a simulated data. It is not from any specific source
}
\keyword{datasets}
