\name{y1}
\alias{y1}
\docType{data}
\title{
y1
}
\description{
a simple vector of 1095 observation. This data set should be used with x1 to see the correlation
}
\usage{data("y1")}
\format{
  The format is:
 num [1:1095] 87.6 80.1 70.1 89.2 65.7 ...
}
\source{
It is a simulated data. It is not from any specific source
}
\keyword{datasets}
