\name{y2}
\alias{y2}
\docType{data}
\title{
y2
}
\description{
a simple vector of 701 observation. This data set should be used with x2 to see the correlation
}
\usage{data("y2")}
\format{
  The format is:
 num [1:701] -3.52 -3.73 -2.3 -3.1 -3.63 ...
}
\source{
It is a simulated data. It is not from any specific source
}
\keyword{datasets}
