\name{y3}
\alias{y3}
\docType{data}
\title{
y3
}
\description{
a simple vector of 4001 observation. This data set should be used with y3 to see the correlation
}
\usage{data("y3")}
\format{
  The format is:
 num [1:4001] 0.0195 0.0159 0.0484 0.0358 0.0149 ...
}
\source{
It is a simulated data. It is not from any specific source
}
\keyword{datasets}
