% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{nlcor}
\alias{nlcor}
\title{Compute Nonlinear Correlation}
\usage{
nlcor(
  x,
  y,
  refine = NA,
  plt = T,
  line_thickness = 1,
  line_opacity = 1,
  chart_title = NA
)
}
\arguments{
\item{x}{A numeric vector. NAs are not allowed. The length of the vector should be more than 10.}

\item{y}{A numeric vector. NAs are not allowed. Length should be same as `x`. The order of
`x` and `y` are relevant. `nlcor(x, y)`` is not equal to `nlcor(y, x)`. Therefore, `x`
should be a causal and `y` should be a dependent variable.}

\item{refine}{Optional. If manually set, a small value, e.g., 0.01, increases
the granularity of local correlation computation. The runtime is faster if `refine` is manually set.
Otherwise, the algorithm automatically finds the best refinement.}

\item{plt}{Optional. Default value FALSE. Set TRUE to return ggplot2 object
for the data correlation visualization.}

\item{line_thickness}{Optional. Default 1. Thickness of the correlation lines. It is a float argument > 0.}

\item{line_opacity}{Optional. Default 1, completely opaque. The opacity of the correlation lines. A float between 0-1. 0 is transparent.}
}
\value{
The output is a list containing the nonlinear correlation \code{cor.estimate}, \code{adjusted.p.value}, and \code{cor.plot}. \code{cor.estimate}
is between 0 and 1 (a negative nonlinear correlation is undefined). The
\code{adjusted.p.value} shows the statistical significance of the
\code{cor.estimate}. If \code{adjusted.p.value > 0.05}, the nonlinear
correlation estimate can be considered as noise (statistically not
significant). If \code{plt = T}, \code{ggplot2} object is return for
plotting the identified local linear correlations in the data.
}
\description{
Compute nonlinear correlation using adaptive spatial sampling.
}
\examples{
library(nlcor)
library(ggplot2)
ncor <- nlcor(x1, y1)
ncor <- nlcor(x2, y2, plt = TRUE)
ncor <- nlcor(x3, y3, refine = 0.01, plt = TRUE)

}
\seealso{
\code{\link{cor}}
}
\keyword{correlation}
\keyword{nonlinear}
