% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{nlcor}
\alias{nlcor}
\title{Compute Nonlinear Correlation}
\usage{
nlcor(x, y)
}
\arguments{
\item{x}{A numeric vector. NAs are not allowed. The length of the vector should be more than 10.}

\item{y}{A numeric vector. NAs are not allowed. Length should be same as x.}
}
\value{
The output is a list containing the nonlinear correlation \code{estimate}.
The estimate is between 0 and 1. An adjusted \code{p.value} is also returned.
}
\description{
Compute nonlinear correlation using spatial sampling. Local linear
correlations are computed at the samples and combined.
}
\examples{
nlcor(x, y)

}
\keyword{correlation}
\keyword{nonlinear}
